refactor : 통합 테스트 중복 코드 제거
