+++ src/test/java/subway/integration/LineIntegrationTest.java
@@ -69,12 +69,12 @@ public class LineIntegrationTest extends IntegrationTest {
                 extract();
 
         // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
     }
 
     @DisplayName("지하철 노선 목록을 조회한다.")
     @Test
-    void getLines() {
+    void getLineList() {
         // given
         ExtractableResponse<Response> createResponse1 = RestAssured
                 .given().log().all()
